💄 ui[clozeText]: Add padding to clozeText for uniform card generation
In this commit, we are updating the parser to improve the look of card generation by adding padding to 'clozeText'. This change was implemented by creating a new context named 'clozeText_padded'. To ensure our cards have a uniform appearance, we added line breaks before and after 'clozeText'. This is being done for both the condition when context is present and when it's not. The process is logged for debugging purposes.